Dev: ui_node: redirect `node delete` to `cluster remove`
It will also adjust the related cluster configurations and clean up the
leftover (eg. stopping the cluster services) on the removed node.

Close #891.
